// src/components/flexbox/types.ts
export enum Direction {
  Row = "Row",
  Column = "Column",
}

export enum JustifyContent {
  start = "START",
  end = "END",
  center = "CENTER",
  between = "BETWEEN",
  around = "AROUND",
  evenly = "EVENLY",
}
export enum AlignItems {
  start = "START",
  end = "END",
  center = "CENTER",
  stretch = "STRETCH",
  baseline = "BASELINE",
}

export enum GapSize {
  small = "4px",
  medium = "8px",
  large = "12px",
  xlarge = "16px",
  xxlarge = "32px",
}